move compute_test_quantities to dict_utils, add docs
